[SETTING] 프로젝트 폴더 및 파일 생성(#2)
- 스토리보드 삭제 -> code base로 설정
