+++ DongSik-e/DongSik-e/Global/Literals/ImageLiteral.swift
@@ -1,0 +1,8 @@
+//
+//  ImageLiteral.swift
+//  DongSik-e
+//
+//  Created by LeeJiSoo on 2023/01/24.
+//
+
+import Foundation
