+++ DongSik-e/DongSik-e/Global/Literals/StringLiteral.swift
@@ -1,0 +1,8 @@
+//
+//  StringLiteral.swift
+//  DongSik-e
+//
+//  Created by LeeJiSoo on 2023/01/24.
+//
+
+import Foundation
