+++ DongSik-e/DongSik-e/Network/Model/SchoolMeal.swift
@@ -1,0 +1,8 @@
+//
+//  SchoolMeal.swift
+//  DongSik-e
+//
+//  Created by LeeJiSoo on 2023/01/24.
+//
+
+import Foundation
